+++ attrs_stubs.go
@@ -1,4 +1,4 @@
-// +build !cgo,!plan9 windows android
+// +build !cgo plan9 windows android
 
 package sftp
