+++ request-plan9.go
@@ -1,0 +1,13 @@
+// +build plan9
+
+package sftp
+
+import "syscall"
+
+func fakeFileInfoSys() interface{} {
+	return &syscall.Dir{}
+}
+
+func testOsSys(sys interface{}) error {
+	return nil
+}
